% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{anocAIC}
\alias{anocAIC}
\title{Comparison of several lmer objects via cAIC}
\usage{
anocAIC(object, ..., digits = 2)
}
\arguments{
\item{object}{a fitted \code{lme4}-object}

\item{...}{additional objects of the same type}

\item{digits}{number of digits to print}
}
\value{
a table comparing the cAIC relevant information of all models
}
\description{
Takes one or more \code{lmer}-objects and produces a table 
to the console.
}
\seealso{
\code{\link{cAIC}} for the model fit.
}
