% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuns_lme.R
\name{family.lme}
\alias{family.lme}
\title{family function for lme objects to have a generic function}
\usage{
\method{family}{lme}(object, ...)
}
\arguments{
\item{object}{lme object}

\item{...}{unused}
}
\value{
returns a Gaussian distribution
}
\description{
family function for lme objects to have a generic function
}
