% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelComponents.R
\name{getModelComponents.lme}
\alias{getModelComponents.lme}
\title{getModelComponents for lme objects}
\usage{
\method{getModelComponents}{lme}(m, analytic = TRUE)
}
\arguments{
\item{m}{lme object}

\item{analytic}{logical}
}
\value{
Model components
}
\description{
getModelComponents for lme objects
}
