% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelComponents.R
\name{getModelComponents.merMod}
\alias{getModelComponents.merMod}
\title{getModelComponents for merMods}
\usage{
\method{getModelComponents}{merMod}(m, analytic)
}
\arguments{
\item{m}{merMod object}

\item{analytic}{logical}
}
\description{
getModelComponents for merMods
}
