% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelComponents.R
\name{getModelComponents}
\alias{getModelComponents}
\title{Generic getModelComponents method}
\usage{
getModelComponents(m, analytic)
}
\arguments{
\item{m}{model object}

\item{analytic}{logical}
}
\value{
Model components
}
\description{
Generic getModelComponents method
}
