% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{guWahbaData}
\alias{guWahbaData}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\references{
Gu and Wahba (1991) Minimizing GCV/GML scores with multiple smoothing parameters 
via the Newton method. SIAM J. Sci. Statist. Comput. 12:383-398
}
\keyword{data}
