% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cAIC}
\alias{print.cAIC}
\title{Print method for cAIC}
\usage{
\method{print}{cAIC}(x, ..., digits = 2)
}
\arguments{
\item{x}{a cAIC object}

\item{...}{further arguments passed to 
generic print function (not in use).}

\item{digits}{number of digits to print}
}
\description{
Print method for cAIC
}
