% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Zambia}
\alias{Zambia}
\title{Subset of the Zambia data set on childhood malnutrition}
\description{
Data analyzed by Kandala et al. (2001) which is used for demonstrative purposes to 
estimate linear mixed and additive models using a stepwise procedure on the basis of the cAIC. 
The full data set is 
available at \url{http://www.uni-goettingen.de/de/551625.html}.
}
\references{
Kandala, N. B., Lang, S., Klasen, S., Fahrmeir, L. (2001): 
Semiparametric Analysis of the Socio-Demographic and Spatial 
Determinants of Undernutrition in Two African Countries. 
Research in Official Statistics, 1, 81-100.
}
\keyword{data}
