% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcondLL.R
\name{getcondLL}
\alias{getcondLL}
\alias{getcondLL.lme}
\alias{getcondLL.merMod}
\title{Function to calculate the conditional log-likelihood}
\usage{
getcondLL(object)

\method{getcondLL}{lme}(object)

\method{getcondLL}{merMod}(object)
}
\arguments{
\item{object}{An object of class \code{merMod} either fitted by
\code{\link[lme4]{lmer}} or \code{\link[lme4]{glmer}} of the 'lme4' package.}
}
\value{
conditional log-likelihood value

\code{NULL}

\code{NULL}
}
\description{
Function to calculate the conditional log-likelihood
}
