% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{guWahbaData}
\alias{guWahbaData}
\title{Data from Gu and Wahba (1991)}
\description{
Data from Gu and Wahba (1991) which is used for demonstrative purposes to exemplarily fit 
a generalized additive mixed model.
}
\references{
Gu and Wahba (1991) Minimizing GCV/GML scores with multiple smoothing parameters 
via the Newton method. SIAM J. Sci. Statist. Comput. 12:383-398
}
\keyword{data}
